LayoutTests/imported/w3c/web-platform-tests/css/css-font-loading/fontface-loading-events.html

<!DOCTYPE html>
<title>Test document.fonts.ready loading with two fonts</title>
<link rel="help" href="https://www.w3.org/TR/css-font-loading/#eventdef-fontfaceset-loadingdone">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.initial {
  font-family: "AhemTest", sans-serif;
  font-size: 20px;
}
</style>
<div>Font loading test</div>
<script>  
  let expectedDoneFace = "AhemTest";
  let expectedErrorFace = "";

  const fontSet = document.fonts;

  promise_test(t => {
    return fontSet.ready;
  }, 'The initial ready promise fired');

  promise_test(t => {
    const eventPromise = new Promise(resolve => {
      fontSet.addEventListener('loadingdone', t.step_func(event => {
        assert_equals(event.fontfaces.length, 1, `'loadingdone' FontFaceSet should have one font; it has ${event.fontfaces.length}`);
        assert_equals(event.fontfaces[0].family, expectedDoneFace, `'loadingdone' FontFaceSet fontFamily is ${expectedDoneFace}`);
        resolve();
      }));
    }, { once: true });

    expectedDoneFace = "AhemTest";
    const fontFace = new FontFace("AhemTest", "url(/fonts/Ahem.ttf)");
    fontSet.add(fontFace);
    fontFace.load();

    return Promise.all([fontSet.ready, eventPromise]);
  }, 'Loading a font should fire a `loadingdone` event with the correct `fontfaceset`');

  promise_test(t => {
    const eventPromise = new Promise(resolve => {
      fontSet.addEventListener('loadingdone', t.step_func(event => {
        assert_equals(event.fontfaces.length, 1, `'loadingdone' FontFaceSet should have one font; it has ${event.fontfaces.length}`);
        assert_equals(event.fontfaces[0].family, expectedDoneFace, `'loadingdone' FontFaceSet fontFamily is ${expectedDoneFace}`);
        resolve();
      }));
    }, { once: true });

    expectedDoneFace = "AhemTest2";
    const fontFace = new FontFace("AhemTest2", "url(/fonts/Ahem.ttf)");
    fontSet.add(fontFace);
    fontFace.load();

    return Promise.all([fontSet.ready, eventPromise]);
  }, 'Loading a second font should fire a `loadingdone` event with the correct `fontfaceset`');

  promise_test(t => {
    const eventPromise = new Promise(resolve => {
      fontSet.addEventListener('loadingerror', t.step_func(event => {
        assert_equals(event.fontfaces.length, 1, `'loadingerror' FontFaceSet should have one font; it has ${event.fontfaces.length}`);
        assert_equals(event.fontfaces[0].family, expectedErrorFace, `'loadingerror' FontFaceSet fontFamily is ${expectedErrorFace}`);
        resolve();
      }));
    }, { once: true });

    expectedErrorFace = "DoesNotExist";
    const fontFace = new FontFace("DoesNotExist", "url(/fonts/DoesNotExist.ttf)");
    fontSet.add(fontFace);
    fontFace.load().catch(() => {});

    return Promise.all([fontSet.ready, eventPromise]);
  }, 'Loading a missing font should fire a `loadingerror` event with the correct `fontfaceset`');

</script>
